+++ src/types.public.ts
@@ -2,6 +2,7 @@
  * General/shared types for this project that are re-exported publicly.
  */
 export {
+    EventsType,
     EventSource,
     EventSourceType,
     EventHandlersType,